requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-bmemcached==0.2.2
django-cors-headers==1.1.0
django-filter==0.9.2
django-phonenumber-field==0.7.2
django-push-notifications==1.2.0
django-storages==1.1.8
django-uuidfield==0.5.0
djangorestframework==3.1.1
djangorestframework-gis==0.8.2
eventlet==0.17.1
gnureadline==6.3.3
greenlet==0.4.5
gunicorn==19.2.1
hashids==1.1.0
httplib2==0.9
httpretty==0.8.8
ipython==2.4.1
jsonfield==1.0.3
mock==1.0.1
phonenumbers==7.0.4
psycopg2==2.6.1
pytz==2014.10
requests==2.5.3
six==1.9.0
static3==0.5.1
twilio==3.7.3
whitenoise==1.0.6
